+++ frontend-react/src/layouts/Markdown/Context.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+const MarkdownLayoutContext = createContext<{
+    sidenavContent?: React.ReactNode;
+    setSidenavContent: (jsx: React.ReactNode) => void;
+    mainContent?: React.ReactNode;
+    setMainContent: (jsx: React.ReactNode) => void;
+}>({} as any);
+
+export default MarkdownLayoutContext;
