--- frontend-react/src/shared/InPageNav/InPageNav.tsx
@@ -1,5 +1,0 @@
-export function InPageNav(props: any) {
-    return <aside className="usa-in-page-nav">{props.children}</aside>;
-}
-
-export default InPageNav;
