--- frontend-react/src/shared/InPageNav/InPageNavList.tsx
@@ -1,5 +1,0 @@
-export function InPageNavList(props: any) {
-    return <ul className="usa-in-page-nav__list">{props.children}</ul>;
-}
-
-export default InPageNavList;
