--- frontend-react/src/shared/InPageNav/InPageNavNav.tsx
@@ -1,5 +1,0 @@
-export function InPageNavNav(props: any) {
-    return <nav className="usa-in-page-nav__nav">{props.children}</nav>;
-}
-
-export default InPageNavNav;
