--- frontend-react/src/shared/InPageNav/index.ts
@@ -1,5 +1,0 @@
-export * from "./InPageNav";
-export * from "./InPageNavNav";
-export * from "./InPageNavHeader";
-export * from "./InPageNavList";
-export * from "./InPageNavItem";
